Move Toml back to `Cli` project
