+++ AuraLang.Cli/AuraLang.Cli.csproj
@@ -11,13 +11,13 @@
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.9.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="Samboy063.Tomlet" Version="5.2.0" />
   </ItemGroup>
   <ItemGroup>
     <None Remove="Commands\" />
     <None Remove="Options\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\AuraLang.Toml\AuraLang.Toml.csproj" />
     <ProjectReference Include="..\AuraLang\AuraLang.csproj" />
   </ItemGroup>
 </Project>
