+++ AuraLang/AuraLang.csproj
@@ -32,7 +32,4 @@
     <Folder Include="Compiler\" />
     <Folder Include="Exceptions\Compiler\" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\AuraLang.Toml\AuraLang.Toml.csproj" />
-  </ItemGroup>
 </Project>
